rsx: Shader pipeline fixes and improvements - Do not set zfunc if alphakill is not enabled. This is because at the moment alphakill requires a different shader to be built
- use glsl loop-unroll friendly comparison; skip vertex input compare if either key requests it

- Minor tweaks to fp key generation

+++ rpcs3/Emu/RSX/RSXFragmentProgram.h
@@ -237,7 +237,7 @@ struct RSXFragmentProgram
 
 	float texture_pitch_scale[16];
 	u8 textures_alpha_kill[16];
-	u32 textures_zfunc[16];
+	u8 textures_zfunc[16];
 
 	bool valid;
